requirements.txt
# Core Dependencies
pandas>=2.0.0,<3.0.0
requests>=2.31.0,<3.0.0
yfinance>=0.2.28,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0

# Database
# SQLite3 is included in Python standard library

# Environment Management
python-dotenv>=1.0.0,<2.0.0

# Data Processing
pyarrow>=14.0.0,<15.0.0  # For Parquet support

# CLI
click>=8.1.0,<9.0.0

# Logging
colorlog>=6.7.0,<7.0.0

# Testing (optional, but recommended)
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.0,<4.0.0

# Development Tools (optional)
black>=23.0.0,<24.0.0
pylint>=3.0.0,<4.0.0
mypy>=1.5.0,<2.0.0

# Type hints
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0

# Rate Limiting (already in existing code)
# No additional package needed - using custom RateLimiter

# Note: Ollama Python client (if needed)
# ollama>=0.1.0,<1.0.0  # Uncomment if using Python client instead of REST API

# LangChain (minimal - polish-only chains)
langchain-core>=0.1.0,<1.0.0
langchain-ollama>=0.1.0,<1.0.0